[Mihut Andrei]-refactoring

+++ News System/src/com/furiapolitehnicii/Listener/Editor.java
@@ -28,11 +28,15 @@ public class Editor implements Listener {
 	}
 
 	public void onNewsAppearEvent(News news) {
+		if(isNewsDeleted(news))
+			return;
 		Event newsAppearEvent = new NewsAppearEvent(news);
 		events.add(newsAppearEvent);
 	}
 
 	public void onNewsDeleteEvent(News news) {
+		if(isNewsDeleted(news))
+			return;
 		news.delete();
 		Event newsAppearEvent = new NewsDeleteEvent(news);
 		events.add(newsAppearEvent);
@@ -40,8 +44,19 @@ public class Editor implements Listener {
 	}
 
 	public void onNewsUpdateEvent(News news) {
+		if(isNewsDeleted(news))
+			return;
 		Event newsAppearEvent = new NewsUpdateEvent(news);
 		events.add(newsAppearEvent);
+	}
+	
+	private static boolean isNewsDeleted(News news)
+	{
+		if (news.isDeleted()) {
+			System.out.println("The news was erased from the editorial. Cannot publish/update/delete a deleted news" + System.lineSeparator() + news.getTitle());
+			return true;
+		}	
+		return false;
 	}
 
 }
